Run queuing enabled endpoint
Summary:
Boolean for whether run queuing is supported by the instance. To be used to show a better null state on the queued runs tab when queuing is not supported, something like

"Run queuing is not supported on this instance. See the QueuedRunCoordinator docs to change this".

Test Plan: Can add a test if desired

Reviewers: dgibson

Reviewed By: dgibson

Subscribers: dish

Differential Revision: https://dagster.phacility.com/D6010