Flip asset materialization UI so newest time/partition is on the left
Summary:
This diff flips the asset materialization UI so it reads newest-to-oldest and the user can see the latest data without scrolling the grid to the far right.

Because we have sparklines beside each row, I flipped them to align with the data being presented in the grid. Since it seemed like the sparklines and the expanded graphs beneath should also match, I flipped those graphs as well so that the newest data is on the left.

I think this is pretty nice overall and we could consider flipping other graphs in Dagit to match. We could alternatively make the grid autoscroll to the far right when it's loaded, but that may not be quite as smooth.

{F503904}

Test Plan: Run tests

Reviewers: schrockn, dish, sandyryza

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D6029