[dagit] QueryCountdown to consolidate some repetitive code
Summary: Now that we've got a few refreshable countdowns in place, it seems like we have a clear pattern of repeated needs that can be abstracted away. Specifically, refreshable countdowns for polling GraphQL queries tend to share the same code, so this diff introduces an Apollo-aware `QueryCountdown` that encapsulates the `useCountdown` hook and network status.

Test Plan: View affected countdown components in Dagit, verify that they behave properly.

Reviewers: bengotow, prha, sashank

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D6324