[execute_in_process improvements] output values for in process execution without reconstructing context
Summary:
This diff adds output tracking for in process execution. This type of output tracking will only work via single process execution / testing cases, because it requires passing the same dictionary to each step context.

This is still a little rough around the edges. Either in the same diff or in a different diff, I'm planning on changing the actual output retrieval mechanism to be more user friendly. Additionally, top-level outputs from a graph are not currently stored (have to go through intermediary of top level solids in dag), so I need to figure that out as well (likely that graphs are split into their constituent solids, will have to add a caching step where the graph outputs are sent to storage or smth.

Test Plan: Unit tests

Reviewers: alangenfeld, sandyryza, yuhan

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D6428