[dagit] Show a tooltip on InstanceWarningIcon
Summary:
In the updated left nav, the Instance warning icon could be a little more informative. This adds a tooltip with a summary of what's prompting us to show the icon at all.

This way, if a user notices it, they can quickly see what's going on and know whether or not to investigate further.

Test Plan: Load Dagit, force a repo error and keep daemons not running. Verify correct tooltip rendering. See screenshots.

Reviewers: bengotow, prha, dgibson, johann

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D6574