fix redis condition in helm requirements.yaml
Summary:
User feedback:
"It appears that you cannot have a negation in a condition in requirements.yaml dependencies and so the redis pods appear to always get started up regardless of whether redis.external==true.
Furthermore, listing a series of comma separated conditions appears to greedily decide based on the first non-undefined value in the list.
https://github.com/helm/helm/issues/4960"

Test Plan: integration

Reviewers: rexledesma

Reviewed By: rexledesma

Differential Revision: https://dagster.phacility.com/D6663