[core] Adding shared Event Log storage test suite
Summary:
This diff adds a shared `TestEventLogStorage` test suite, much like we have for Run Storages & Schedule storages. Some tests have to be skipped because they rely on implementation-specific behavior (i.e. a backing SQL database) but I've tried to refactor so that as many tests are shared as possible.

The TestEventLogStorage suite is the combination of tests we formerly had for `InMemory`, `Sqlite`, `ConsolidatedSqlite`, and `Postgres` `EventLogStorage`s

The motivation for this diff is a comment from Phil in D6617

Test Plan: this is a refactoring of the event log test suite

Reviewers: catherinewu, rexledesma, prha, max

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D6701