Tweak spacing and items in simple left nav
Summary:
Make the hit area a little bit bigger for the left nav items in the simplified nav.

Also remove "Solids", which it seems like we may not really want, and remove the focus outline on the warning icon since it's not really meant to be individually clickable.

Test Plan: Toggle new nav on, view Dagit. Verify that everything looks fine.

Reviewers: bengotow, sandyryza

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D6746