[dagit] Separate pipeline and snapshot links in Run header
Summary:
Based on some feedback from schrockn and sandyryza in a conversation yesterday, split the "pipeline@snapshot" link in the Run root header to be two links:

- The pipeline name to `/workspace/pipelines/{pipelineName}`, which may go to a "not found" page, depending on which workspaces are loaded)
- The snapshot ID to `/instance/snapshots/{snapshot}`, which will always go to a snapshot page.

It's a bit of a subtle change, but hopefully helps save a click in 99% of use cases.

Test Plan: View a run, verify that the links work as described above.

Reviewers: bengotow, schrockn, sandyryza

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D6792