un-config-itize longitudinal pipeline
Summary:
This shifts the reusable logic in the longitudinal pipeline from a config pattern to a solid factory pattern.  This will make it easier to layer on a couple additions:
* Adding a bunch of metadata so we'll have a realistic fleshed out example that we can include in screenshots
* Asset lineage

This is maybe a subjective take, but I think it's also a nice demonstration of how things get simpler and require less code when there's less reliance on config.

Test Plan: ran a backfill in local dagit

Reviewers: owen, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D6817