better build_resources context object
Summary:
This diff creates a named tuple scoping the resources provided to the `build_resources` context manager.

The main thing I was deciding between was a namedtuple approach and using a frozendict directly. Ultimately, using a namedtuple approach more closely matches the experience of the user when executing a solid, and the less conflicting APIs the better.

The only downside is we have to use getattr internally when using this API, but that isn't too bad.

Test Plan: Existing tests.

Reviewers: alangenfeld, sandyryza, schrockn

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D6875