Remove instance creation in sensor/schedule gRPC
Summary:
User code containers do not write to the db unless a run is launched via Dagit or Daemon. However, currently they read from the db since the instance is constructed when Dagit/Daemon asks for schedule or sensor execution. This should exist solely in Dagit/Daemon, so we defer the creation of the instance unless the user explicitly needs it in their sensor/schedule business logic.

With this change, the db readiness check only should exist in Dagit in Daemon, rather than the user code container.

Test Plan:
pytest
integration

Reviewers: dgibson, catherinewu, johann, prha

Reviewed By: dgibson, prha

Differential Revision: https://dagster.phacility.com/D6883